Moved is_external_repository to get_content_path
