+++ demisto_sdk/commands/common/content_constant_paths.py
@@ -1,7 +1,7 @@
 import logging
 from pathlib import Path
 
-from demisto_sdk.commands.common.tools import get_content_path
+from demisto_sdk.commands.common.get_content_path import get_content_path
 
 logger = logging.getLogger("demisto-sdk")
