+++ demisto_sdk/commands/common/hook_validations/pack_unique_files.py
@@ -45,6 +45,7 @@ from demisto_sdk.commands.common.constants import (  # PACK_METADATA_PRICE,
 from demisto_sdk.commands.common.content import Content
 from demisto_sdk.commands.common.content.objects.pack_objects.pack import Pack
 from demisto_sdk.commands.common.errors import Errors
+from demisto_sdk.commands.common.get_content_path import is_external_repository
 from demisto_sdk.commands.common.git_util import GitUtil
 from demisto_sdk.commands.common.handlers import DEFAULT_JSON_HANDLER as json
 from demisto_sdk.commands.common.hook_validations.base_validator import (
@@ -756,7 +757,7 @@ class PackUniqueFilesValidator(BaseValidator):
         Return:
              bool: True if the usecases are approved, otherwise False
         """
-        if tools.is_external_repository():
+        if is_external_repository():
             return True
 
         non_approved_usecases = set()
@@ -803,7 +804,7 @@ class PackUniqueFilesValidator(BaseValidator):
         Return:
             bool: True if the tags are approved, otherwise False
         """
-        if tools.is_external_repository():
+        if is_external_repository():
             return True
 
         is_valid = True
@@ -831,7 +832,7 @@ class PackUniqueFilesValidator(BaseValidator):
         Return:
              bool: True if the tags are approved, otherwise False
         """
-        if tools.is_external_repository():
+        if is_external_repository():
             return True
 
         is_valid_tag_prefixes = True
@@ -1172,7 +1173,7 @@ class PackUniqueFilesValidator(BaseValidator):
         Returns:
             bool: True if pack contain only one category and the category is from the approved list. Otherwise, return False.
         """
-        if tools.is_external_repository():
+        if is_external_repository():
             return True
         categories = self._read_metadata_content().get("categories", [])
         approved_list = tools.get_current_categories()
