+++ demisto_sdk/commands/common/tests/pack_unique_files_test.py
@@ -8,7 +8,7 @@ from click.testing import CliRunner
 from git import GitCommandError
 
 from demisto_sdk.__main__ import main
-from demisto_sdk.commands.common import tools
+from demisto_sdk.commands.common import get_content_path, tools
 from demisto_sdk.commands.common.constants import (
     PACK_METADATA_DESC,
     PACK_METADATA_SUPPORT,
@@ -189,7 +189,9 @@ class TestPackUniqueFilesValidator:
         mocker.patch.object(ValidateManager, "setup_git_params", return_value=True)
         pack_metadata_no_email_and_url["email"] = ""
         pack_metadata_no_email_and_url["url"] = ""
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(
             PackUniqueFilesValidator, "_is_pack_file_exists", return_value=True
         )
@@ -246,7 +248,9 @@ class TestPackUniqueFilesValidator:
 
         pack_metadata_changed_url = PACK_METADATA_PARTNER.copy()
         pack_metadata_changed_url["url"] = url
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(
             PackUniqueFilesValidator, "_is_pack_file_exists", return_value=True
         )
@@ -300,7 +304,9 @@ class TestPackUniqueFilesValidator:
 
         pack_metadata_price_changed = PACK_METADATA_PARTNER.copy()
         pack_metadata_price_changed["price"] = 3
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(
             PackUniqueFilesValidator, "_is_pack_file_exists", return_value=True
         )
@@ -477,7 +483,9 @@ class TestPackUniqueFilesValidator:
                 PACK_METADATA_TAGS: [],
             }
         )
-        mocker.patch.object(tools, "is_external_repository", return_value=False)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=False
+        )
         mocker.patch.object(
             tools,
             "get_dict_from_file",
@@ -564,7 +572,9 @@ class TestPackUniqueFilesValidator:
                 PACK_METADATA_TAGS: tags,
             }
         )
-        mocker.patch.object(tools, "is_external_repository", return_value=False)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=False
+        )
         mocker.patch.object(
             tools,
             "get_dict_from_file",
@@ -612,7 +622,9 @@ class TestPackUniqueFilesValidator:
             }
         )
 
-        mocker.patch.object(tools, "is_external_repository", return_value=False)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=False
+        )
         mocker.patch.object(
             tools,
             "get_dict_from_file",
@@ -649,7 +661,9 @@ class TestPackUniqueFilesValidator:
                 PACK_METADATA_TAGS: tags,
             }
         )
-        mocker.patch.object(tools, "is_external_repository", return_value=False)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=False
+        )
         self.validator.pack_path = pack.path
 
         with ChangeCWD(repo.path):
