+++ demisto_sdk/commands/common/tools.py
@@ -114,6 +114,10 @@ from demisto_sdk.commands.common.constants import (
     urljoin,
 )
 from demisto_sdk.commands.common.git_content_config import GitContentConfig, GitProvider
+from demisto_sdk.commands.common.get_content_path import (
+    get_content_path,
+    is_external_repository,
+)
 from demisto_sdk.commands.common.git_util import GitUtil
 from demisto_sdk.commands.common.handlers import DEFAULT_JSON_HANDLER as json
 from demisto_sdk.commands.common.handlers import DEFAULT_YAML_HANDLER as yaml
@@ -1970,19 +1974,6 @@ def get_common_server_dir_pwsh(env_dir):
     return _get_common_server_dir_general(env_dir, "CommonServerPowerShell")
 
 
-def is_external_repository() -> bool:
-    """
-    Returns True if script executed from private repository
-
-    """
-    try:
-        git_repo = git.Repo(os.getcwd(), search_parent_directories=True)
-        private_settings_path = os.path.join(git_repo.working_dir, ".private-repo-settings")  # type: ignore
-        return os.path.exists(private_settings_path)
-    except git.InvalidGitRepositoryError:
-        return True
-
-
 def get_content_id_set() -> dict:
     """Getting the ID Set from official content's bucket"""
     return requests.get(OFFICIAL_CONTENT_ID_SET_PATH).json()
@@ -2015,35 +2006,6 @@ def get_latest_upload_flow_commit_hash() -> str:
     if not last_commit:
         raise ValueError("The latest commit hash was not found in the index.json file")
     return last_commit
-
-
-def get_content_path() -> Path:
-    """Get abs content path, from any CWD
-    Returns:
-        str: Absolute content path
-    """
-    try:
-        if content_path := os.getenv("DEMISTO_SDK_CONTENT_PATH"):
-            git_repo = git.Repo(content_path)
-            logger.debug(f"Using content path: {content_path}")
-        else:
-            git_repo = git.Repo(Path.cwd(), search_parent_directories=True)
-
-        remote_url = git_repo.remote().urls.__next__()
-        is_fork_repo = "content" in remote_url
-        is_external_repo = is_external_repository()
-
-        if not is_fork_repo and not is_external_repo:
-            raise git.InvalidGitRepositoryError
-        if not git_repo.working_dir:
-            return Path.cwd()
-        return Path(git_repo.working_dir)
-    except (git.InvalidGitRepositoryError, git.NoSuchPathError):
-        if not os.getenv("DEMISTO_SDK_IGNORE_CONTENT_WARNING"):
-            logger.info(
-                "[yellow]Please run demisto-sdk in content repository![/yellow]"
-            )
-    return Path(".")
 
 
 def run_command_os(
