+++ demisto_sdk/tests/integration_tests/validate_integration_test.py
@@ -11,7 +11,7 @@ from requests.adapters import HTTPAdapter
 from urllib3 import Retry
 
 from demisto_sdk.__main__ import main
-from demisto_sdk.commands.common import tools
+from demisto_sdk.commands.common import get_content_path
 from demisto_sdk.commands.common.constants import DEFAULT_IMAGE_BASE64
 from demisto_sdk.commands.common.content.content import Content
 from demisto_sdk.commands.common.git_util import GitUtil
@@ -145,7 +145,9 @@ class TestGenericFieldValidation:
         - Ensure success validation message is printed.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         pack.create_generic_field("generic-field", GENERIC_FIELD)
         generic_field_path = pack.generic_fields[0].path
@@ -177,7 +179,9 @@ class TestGenericFieldValidation:
         - Ensure validation fails on ST108 - a field which doesn't defined in the scheme.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         generic_field_copy = GENERIC_FIELD.copy()
         generic_field_copy["test"] = True
@@ -222,7 +226,9 @@ class TestGenericFieldValidation:
         - Ensure validation fails with the right error code.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         generic_field_copy = GENERIC_FIELD.copy()
         generic_field_copy[field_to_test] = invalid_value
@@ -261,7 +267,9 @@ class TestGenericTypeValidation:
         - Ensure success validation message is printed.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         pack.create_generic_type("generic-type", GENERIC_TYPE)
         generic_type_path = pack.generic_types[0].path
@@ -293,7 +301,9 @@ class TestGenericTypeValidation:
         - Ensure validation fails on ST108 - a field which doesn't defined in the scheme.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         generic_type_copy = GENERIC_TYPE.copy()
         generic_type_copy["test"] = True
@@ -328,7 +338,9 @@ class TestGenericTypeValidation:
         - Ensure validation fails on BA106 - no minimal fromversion in file.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         generic_type_copy = GENERIC_TYPE.copy()
         generic_type_copy["fromVersion"] = "6.0.0"
@@ -366,7 +378,9 @@ class TestGenericModuleValidation:
         - Ensure success validation message is printed.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         pack.create_generic_module("generic-module", GENERIC_MODULE)
         generic_module_path = pack.generic_modules[0].path
@@ -398,7 +412,9 @@ class TestGenericModuleValidation:
         - Ensure validation fails on ST108 - a field which doesn't defined in the scheme.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         generic_module_copy = GENERIC_MODULE.copy()
         generic_module_copy["test"] = True
@@ -433,7 +449,9 @@ class TestGenericModuleValidation:
         - Ensure validation fails on BA106 - no minimal fromversion in file.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         generic_module_copy = GENERIC_MODULE.copy()
         generic_module_copy["fromVersion"] = "6.0.0"
@@ -471,7 +489,9 @@ class TestGenericDefinitionValidation:
         - Ensure success validation message is printed.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         generic_def_copy = GENERIC_DEFINITION.copy()
         genefic_def = pack.create_generic_definition(
@@ -505,7 +525,9 @@ class TestGenericDefinitionValidation:
         - Ensure validation fails.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         generic_def_copy = GENERIC_DEFINITION.copy()
         generic_def_copy["anotherField"] = False
@@ -541,7 +563,9 @@ class TestGenericDefinitionValidation:
         - Ensure validation fails on BA106 - no minimal fromversion in file.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         generic_def_copy = GENERIC_DEFINITION.copy()
         generic_def_copy["fromVersion"] = "6.0.0"
@@ -580,7 +604,9 @@ class TestIncidentFieldValidation:
         - Ensure success validation message is printed.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         pack.create_incident_field("incident-field", INCIDENT_FIELD)
         incident_field_path = pack.incident_fields[0].path
@@ -613,7 +639,9 @@ class TestIncidentFieldValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         incident_field_copy = INCIDENT_FIELD.copy()
         incident_field_copy["system"] = True
@@ -651,7 +679,9 @@ class TestIncidentFieldValidation:
         - Ensure success validation message is printed.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         incident_field_copy = INCIDENT_FIELD.copy()
         incident_field_copy["script"] = "test_script"
@@ -721,7 +751,9 @@ class TestIncidentFieldValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         incident_field_copy = INCIDENT_FIELD.copy()
         incident_field_copy["script"] = "test_script"
@@ -786,7 +818,9 @@ class TestDeprecatedIntegration:
         - Ensure validation passes.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         pack = repo.create_pack("PackName")
 
@@ -830,7 +864,9 @@ class TestDeprecatedIntegration:
         - Ensure validation fails on - invalid_deprecated_integration.
         """
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         pack = repo.create_pack("PackName")
         invalid_integration_yml = deepcopy(
@@ -866,7 +902,9 @@ class TestDeprecatedIntegration:
                 Ensure validation fails
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         pack = repo.create_pack("PackName")
 
@@ -906,7 +944,9 @@ class TestDeprecatedIntegration:
                 Ensure validation fails
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         pack = repo.create_pack("PackName")
 
@@ -940,7 +980,9 @@ class TestDeprecatedIntegration:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         pack = repo.create_pack("PackName")
         pack_integration_path = join(
@@ -982,7 +1024,9 @@ class TestDeprecatedIntegration:
         - Ensure validation passes.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         pack = repo.create_pack("PackName")
         pack_integration_path = join(
@@ -1033,7 +1077,9 @@ class TestDeprecatedIntegration:
         - Ensure validation passes.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         mocker.patch.object(
             PackUniqueFilesValidator, "are_valid_files", return_value=""
@@ -1106,7 +1152,9 @@ class TestDeprecatedIntegration:
         - Ensure validation passes.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         pack = repo.create_pack("PackName")
         pack_integration_path = join(
@@ -1153,7 +1201,9 @@ class TestDeprecatedIntegration:
         - Ensure validation passes.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         mocker.patch.object(ValidateManager, "setup_git_params", return_value=True)
         mocker.patch.object(
@@ -1224,7 +1274,9 @@ class TestIntegrationValidation:
         - Ensure validate passes and identifies the file as an integration.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         pack_integration_path = join(
             AZURE_FEED_PACK_PATH, "Integrations/FeedAzure/FeedAzure.yml"
@@ -1265,7 +1317,9 @@ class TestIntegrationValidation:
         - Ensure validate fails on wrong required value
         """
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(
             PackUniqueFilesValidator, "are_valid_files", return_value=""
         )
@@ -1341,7 +1395,9 @@ class TestIntegrationValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         mocker.patch.object(
             IntegrationValidator, "is_valid_category", return_value=True
@@ -1514,7 +1570,9 @@ class TestIntegrationValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         mocker.patch.object(
             IntegrationValidator, "is_valid_category", return_value=True
@@ -1572,7 +1630,9 @@ class TestIntegrationValidation:
         - Ensure validate fails on ST107 - pykwalify_missing_parameter.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         pack = repo.create_pack("PackName")
         pack_integration_path = join(
@@ -1791,7 +1851,9 @@ class TestClassifierValidation:
         - Ensure validate passes.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         classifier = pack.create_classifier("new_classifier", NEW_CLASSIFIER)
         with ChangeCWD(pack.repo_path):
@@ -1823,7 +1885,9 @@ class TestClassifierValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         new_classifier_copy = NEW_CLASSIFIER.copy()
         new_classifier_copy["fromVersion"] = "5.0.0"
@@ -1861,7 +1925,9 @@ class TestClassifierValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         new_classifier_copy = NEW_CLASSIFIER.copy()
         new_classifier_copy["toVersion"] = "5.0.0"
@@ -1893,7 +1959,9 @@ class TestClassifierValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         new_classifier_copy = NEW_CLASSIFIER.copy()
         new_classifier_copy["toVersion"] = "6.0.2"
@@ -1926,7 +1994,9 @@ class TestClassifierValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         new_classifier_copy = NEW_CLASSIFIER.copy()
         del new_classifier_copy["id"]
@@ -1962,7 +2032,9 @@ class TestClassifierValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         new_classifier_copy = NEW_CLASSIFIER.copy()
         del new_classifier_copy["fromVersion"]
@@ -1997,7 +2069,9 @@ class TestClassifierValidation:
         - Ensure validate found errors.
         """
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         new_classifier_copy = NEW_CLASSIFIER.copy()
         new_classifier_copy["type"] = "test"
@@ -2025,7 +2099,9 @@ class TestClassifierValidation:
         - Ensure validate passes.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(
             ClassifierValidator, "is_incident_field_exist", return_value=True
         )
@@ -2060,7 +2136,9 @@ class TestClassifierValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         old_classifier_copy = OLD_CLASSIFIER.copy()
         old_classifier_copy["fromVersion"] = "6.0.0"
@@ -2095,7 +2173,9 @@ class TestClassifierValidation:
         - Ensure validate found errors.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         old_classifier_copy = OLD_CLASSIFIER.copy()
         old_classifier_copy["toVersion"] = "6.0.0"
@@ -2129,7 +2209,9 @@ class TestClassifierValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         old_classifier_copy = OLD_CLASSIFIER.copy()
         del old_classifier_copy["id"]
@@ -2165,7 +2247,9 @@ class TestClassifierValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         old_classifier_copy = OLD_CLASSIFIER.copy()
         del old_classifier_copy["toVersion"]
@@ -2202,7 +2286,9 @@ class TestMapperValidation:
         - Ensure validate passes.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(
             MapperValidator, "is_incident_field_exist", return_value=True
         )
@@ -2237,7 +2323,9 @@ class TestMapperValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         mapper_copy = MAPPER.copy()
         mapper_copy["fromVersion"] = "5.0.0"
@@ -2274,7 +2362,9 @@ class TestMapperValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         mapper_copy = MAPPER.copy()
         mapper_copy["toVersion"] = "5.0.0"
@@ -2306,7 +2396,9 @@ class TestMapperValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         mapper_copy = MAPPER.copy()
         del mapper_copy["id"]
@@ -2342,7 +2434,9 @@ class TestMapperValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         mapper_copy = MAPPER.copy()
         mapper_copy["toVersion"] = "6.0.2"
@@ -2374,7 +2468,9 @@ class TestMapperValidation:
         - Ensure validate found errors.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         mapper_copy = MAPPER.copy()
         mapper_copy["type"] = "test"
@@ -2404,7 +2500,9 @@ class TestDashboardValidation:
         - Ensure validate passes and identifies the file as a dashboard.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         dashboard = pack.create_dashboard("dashboard", DASHBOARD)
         with ChangeCWD(pack.repo_path):
@@ -2436,7 +2534,9 @@ class TestDashboardValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         dashboard_copy = DASHBOARD.copy()
         dashboard_copy["version"] = 1
@@ -2475,7 +2575,9 @@ class TestConnectionValidation:
         - Ensure validate passes and identifies the file as a connection.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         connection = pack._create_json_based(
             name="connection", prefix="", content=CONNECTION
@@ -2509,7 +2611,9 @@ class TestConnectionValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         connection_copy = CONNECTION.copy()
         del connection_copy["canvasContextConnections"][0]["contextKey1"]
@@ -2548,7 +2652,9 @@ class TestIndicatorFieldValidation:
         - Ensure validate passes and identifies the file as an indicator field.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         pack.create_indicator_field("indicator-field", INDICATOR_FIELD)
         indicator_field_path = pack.indicator_fields[0].path
@@ -2581,7 +2687,9 @@ class TestIndicatorFieldValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         indicator_field_copy = INDICATOR_FIELD.copy()
         indicator_field_copy["content"] = False
@@ -2620,7 +2728,9 @@ class TestIncidentTypeValidation:
         - Ensure validate passes and identifies the file as an incident type.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         incident_type = pack.create_incident_type("incident_type", INCIDENT_TYPE)
         with ChangeCWD(pack.repo_path):
@@ -2652,7 +2762,9 @@ class TestIncidentTypeValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         incident_type_copy = INCIDENT_TYPE.copy()
         incident_type_copy["days"] = -1
@@ -2689,7 +2801,9 @@ class TestIncidentTypeValidation:
         - Ensure validate passes and identifies the file as an incident type.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         incident_type_data = INCIDENT_TYPE.copy()
         incident_type_data["extractSettings"] = {
@@ -2751,7 +2865,9 @@ class TestIncidentTypeValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         incident_type_data = INCIDENT_TYPE.copy()
         incident_type_data["extractSettings"] = {
@@ -2817,7 +2933,9 @@ class TestIncidentTypeValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         incident_type_data = INCIDENT_TYPE.copy()
         incident_type_data["extractSettings"] = {
@@ -2936,7 +3054,9 @@ class TestLayoutValidation:
         - Ensure validate passes and identifies the file as a layout.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         layout = pack._create_json_based(name="layout-name", prefix="", content=LAYOUT)
         with ChangeCWD(pack.repo_path):
@@ -2968,7 +3088,9 @@ class TestLayoutValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         layout_copy = LAYOUT.copy()
         layout_copy["version"] = 2
@@ -3008,7 +3130,9 @@ class TestLayoutValidation:
         """
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         layout_copy = LAYOUT.copy()
         layout_copy["version"] = 2
@@ -3046,7 +3170,9 @@ class TestLayoutValidation:
         - Ensure validate passes and identifies the file as a layout.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         layout = pack._create_json_based(
             name="layoutscontainer-test", prefix="", content=LAYOUTS_CONTAINER
@@ -3080,7 +3206,9 @@ class TestLayoutValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         layout_copy = LAYOUTS_CONTAINER.copy()
         layout_copy["version"] = 2
@@ -3120,7 +3248,9 @@ class TestLayoutValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         layout = pack._create_json_based(
             name="wrongname", prefix="", content=LAYOUTS_CONTAINER
@@ -3157,7 +3287,9 @@ class TestLayoutValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         layoutscontainer_copy = LAYOUTS_CONTAINER.copy()
         layoutscontainer_copy["fromVersion"] = "5.0.0"
@@ -3196,7 +3328,9 @@ class TestLayoutValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         layout_copy = LAYOUT.copy()
         layout_copy["toVersion"] = "6.0.0"
@@ -3234,7 +3368,9 @@ class TestLayoutValidation:
         - Ensure success validation message is printed.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         layoutscontainer_copy = LAYOUTS_CONTAINER.copy()
         layoutscontainer_copy["detailsV2"]["tabs"][0]["sections"] = [
@@ -3304,7 +3440,9 @@ class TestLayoutValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         layoutscontainer_copy = LAYOUTS_CONTAINER.copy()
         layoutscontainer_copy["detailsV2"]["tabs"][0]["sections"] = [
@@ -3375,7 +3513,9 @@ class TestLayoutValidation:
         - Ensure success validation message is printed.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         layout_copy = LAYOUT.copy()
         layout_copy["layout"]["tabs"][0]["sections"] = [tab_section_to_test]
@@ -3440,7 +3580,9 @@ class TestLayoutValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         layout_copy = LAYOUT.copy()
         layout_copy["layout"]["tabs"][0]["sections"] = [tab_section_to_test]
@@ -3505,7 +3647,9 @@ class TestPlaybookValidation:
         - Ensure validate passes and identifies the file as a playbook.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(PlaybookValidator, "is_script_id_valid", return_value=True)
         mocker.patch.object(
             ContentEntityValidator, "validate_readme_exists", return_value=True
@@ -3546,7 +3690,9 @@ class TestPlaybookValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         with ChangeCWD(TEST_FILES_PATH):
             runner = CliRunner(mix_stderr=False)
             result = runner.invoke(
@@ -3588,7 +3734,9 @@ class TestPlaybookValidateDeprecated:
         - Ensure validate passes and identifies the file as a playbook deprecated.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(PlaybookValidator, "is_script_id_valid", return_value=True)
         runner = CliRunner(mix_stderr=False)
         result = runner.invoke(
@@ -3626,7 +3774,9 @@ class TestPlaybookValidateDeprecated:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         with ChangeCWD(TEST_FILES_PATH):
             runner = CliRunner(mix_stderr=False)
             result = runner.invoke(
@@ -3662,7 +3812,9 @@ class TestPlaybookValidateDeprecated:
         - Ensure validate passes and identifies the file as a playbook deprecated.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(PlaybookValidator, "is_script_id_valid", return_value=True)
         pack = repo.create_pack("PackName")
         valid_playbook_yml = get_yaml(VALID_DEPRECATED_PLAYBOOK_FILE_PATH)
@@ -3699,7 +3851,9 @@ class TestPlaybookValidateDeprecated:
         - Ensure validate passes and identifies the file as a playbook deprecated.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(PlaybookValidator, "is_script_id_valid", return_value=True)
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         mocker.patch.object(
@@ -3763,7 +3917,9 @@ class TestPlaybookValidateDeprecated:
         - Ensure validate passes and identifies the file as a playbook deprecated.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(PlaybookValidator, "is_script_id_valid", return_value=True)
         pack = repo.create_pack("PackName")
         valid_playbook_yml = get_yaml(VALID_DEPRECATED_PLAYBOOK_FILE_PATH)
@@ -3801,7 +3957,9 @@ class TestPlaybookValidateDeprecated:
         - Ensure validate passes and identifies the file as a playbook deprecated.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(PlaybookValidator, "is_script_id_valid", return_value=True)
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         mocker.patch.object(
@@ -3866,7 +4024,9 @@ class TestReportValidation:
         - Ensure validate passes and identifies the file as a report.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         report = pack._create_json_based(name="report", prefix="", content=REPORT)
         with ChangeCWD(pack.repo_path):
@@ -3898,7 +4058,9 @@ class TestReportValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         report_copy = REPORT.copy()
         report_copy["orientation"] = "bla"
@@ -3930,7 +4092,9 @@ class TestReputationValidation:
         - Ensure validate passes and identifies the file as a reputation.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         reputation = pack._create_json_based(
             name="reputation", prefix="", content=REPUTATION
@@ -3964,7 +4128,9 @@ class TestReputationValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         reputation_copy = REPUTATION.copy()
         reputation_copy["expiration"] = -1
@@ -4002,7 +4168,9 @@ class TestScriptValidation:
         - Ensure validate passes and identifies the file as a script.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         valid_script_yml = get_yaml(VALID_SCRIPT_PATH)
         script = pack.create_script(yml=valid_script_yml)
@@ -4037,7 +4205,9 @@ class TestScriptValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         pack = repo.create_pack("PackName")
         invalid_script_yml = get_yaml(VALID_SCRIPT_PATH)
@@ -4078,7 +4248,9 @@ class TestScriptDeprecatedValidation:
         - Ensure validate passes and identifies the file as a deprecated script.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         valid_script_yml = get_yaml(VALID_SCRIPT_PATH)
         valid_script_yml["deprecated"] = True
@@ -4117,7 +4289,9 @@ class TestScriptDeprecatedValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         pack = repo.create_pack("PackName")
         invalid_script_yml = get_yaml(VALID_SCRIPT_PATH, cache_clear=True)
@@ -4154,7 +4328,9 @@ class TestScriptDeprecatedValidation:
         - Ensure validate passes and identifies the file as a deprecated script.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         valid_script_yml = get_yaml(VALID_SCRIPT_PATH)
         valid_script_yml["deprecated"] = True
@@ -4199,7 +4375,9 @@ class TestScriptDeprecatedValidation:
         - Ensure validate passes and identifies the file as a deprecated script.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         mocker.patch.object(
             PackUniqueFilesValidator, "are_valid_files", return_value=""
@@ -4268,7 +4446,9 @@ class TestScriptDeprecatedValidation:
         - Ensure validate passes and identifies the file as a script.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         valid_script_yml = get_yaml(VALID_SCRIPT_PATH)
         valid_script_yml["toversion"] = "4.4.4"
@@ -4310,7 +4490,9 @@ class TestScriptDeprecatedValidation:
         - Ensure validate passes and identifies the file as a script.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         mocker.patch.object(
             PackUniqueFilesValidator, "are_valid_files", return_value=""
@@ -4376,7 +4558,9 @@ class TestWidgetValidation:
         - Ensure validate passes and identifies the file as a widget.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         widget = pack._create_json_based(name="widget", prefix="", content=WIDGET)
         with ChangeCWD(pack.repo_path):
@@ -4408,7 +4592,9 @@ class TestWidgetValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         widget_copy = WIDGET.copy()
         widget_copy["version"] = 1
@@ -4447,7 +4633,9 @@ class TestImageValidation:
         - Ensure validate passes and identifies the file as an image.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         integration = pack.create_integration()
         image_path = integration.image.path
@@ -4480,7 +4668,9 @@ class TestImageValidation:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         pack = repo.create_pack("PackName")
         integration = pack.create_integration()
@@ -4521,7 +4711,9 @@ class TestImageValidation:
         - Ensure validate does not validates it as an image.
         """
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         pack = repo.create_pack("PackName")
         with ChangeCWD(pack.repo_path):
@@ -4591,7 +4783,9 @@ class TestAuthorImageValidation:
         - Ensure validate passes.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         pack.pack_metadata.write_json(
             {
@@ -4641,7 +4835,9 @@ class TestAuthorImageValidation:
         - Ensure validate fails on error IM108 - empty author image error.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(ImageValidator, "load_image", return_value="")
         pack = repo.create_pack("PackName")
         pack.pack_metadata.write_json(
@@ -4693,7 +4889,9 @@ class TestAllFilesValidator:
         - Ensure validate passes on all files.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=False)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=False
+        )
         mocker.patch.object(
             PackUniqueFilesValidator, "are_valid_files", return_value=""
         )
@@ -4769,7 +4967,9 @@ class TestAllFilesValidator:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=False)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=False
+        )
         mocker.patch.object(
             PackUniqueFilesValidator, "are_valid_files", return_value=""
         )
@@ -4852,7 +5052,9 @@ class TestValidationUsingGit:
         - Ensure validate passes on all files.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=False)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=False
+        )
         mocker.patch.object(
             PackUniqueFilesValidator, "are_valid_files", return_value=""
         )
@@ -4954,7 +5156,9 @@ class TestValidationUsingGit:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=False)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=False
+        )
         mocker.patch.object(
             PackUniqueFilesValidator, "are_valid_files", return_value=""
         )
@@ -5059,7 +5263,9 @@ class TestValidationUsingGit:
             ),
         )
         modified_files = {integration.yml.rel_path}
-        mocker.patch.object(tools, "is_external_repository", return_value=False)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=False
+        )
         mocker.patch.object(
             BaseValidator, "update_checked_flags_by_support_level", return_value=None
         )
@@ -5134,7 +5340,9 @@ class TestValidationUsingGit:
             ),
         )
         modified_files = {integration.yml.rel_path}
-        mocker.patch.object(tools, "is_external_repository", return_value=False)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=False
+        )
         mocker.patch.object(ValidateManager, "setup_git_params", return_value=True)
         mocker.patch.object(
             ValidateManager, "setup_prev_ver", return_value="origin/master"
@@ -5223,7 +5431,9 @@ class TestValidationUsingGit:
         - Ensure an error is raised on the non found file
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=False)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=False
+        )
         mocker.patch.object(ValidateManager, "setup_git_params", return_value=True)
         mocker.patch.object(PackDependencies, "find_dependencies", return_value={})
         mocker.patch.object(
@@ -5280,7 +5490,9 @@ class TestValidationUsingGit:
         script.create_default_script()
 
         modified_files = {integration.yml.rel_path, script.yml.rel_path}
-        mocker.patch.object(tools, "is_external_repository", return_value=False)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=False
+        )
         mocker.patch.object(ValidateManager, "setup_git_params", return_value=True)
         mocker.patch.object(
             ValidateManager, "setup_prev_ver", return_value="origin/master"
@@ -5357,7 +5569,9 @@ class TestValidationUsingGit:
             (integration.yml.rel_path, integration.yml.rel_path),
             script.yml.rel_path,
         }
-        mocker.patch.object(tools, "is_external_repository", return_value=False)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=False
+        )
         mocker.patch.object(ValidateManager, "setup_git_params", return_value=True)
         mocker.patch.object(
             ValidateManager, "setup_prev_ver", return_value="origin/master"
@@ -5442,7 +5656,9 @@ class TestValidationUsingGit:
             integration_2.yml.rel_path,
             script_2.yml.rel_path,
         }
-        mocker.patch.object(tools, "is_external_repository", return_value=False)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=False
+        )
         mocker.patch.object(ValidateManager, "setup_git_params", return_value=True)
         mocker.patch.object(
             ValidateManager, "setup_prev_ver", return_value="origin/master"
@@ -5518,7 +5734,9 @@ class TestSpecificValidations:
         due to the flag.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         reputation_copy = REPUTATION.copy()
         reputation_copy["expiration"] = -1
@@ -5562,7 +5780,9 @@ class TestSpecificValidations:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         reputation_copy = REPUTATION.copy()
         reputation_copy["expiration"] = -1
@@ -5612,7 +5832,9 @@ class TestSpecificValidations:
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
         logger_error = mocker.patch.object(logging.getLogger("demisto-sdk"), "error")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         pack = repo.create_pack("PackName")
         reputation_copy = REPUTATION.copy()
         reputation_copy["expiration"] = -1
@@ -5663,7 +5885,9 @@ class TestBasicValidation:
         - make sure the files are valid and that the errors can be ignored successfully.
         """
         logger_info = mocker.patch.object(logging.getLogger("demisto-sdk"), "info")
-        mocker.patch.object(tools, "is_external_repository", return_value=True)
+        mocker.patch.object(
+            get_content_path, "is_external_repository", return_value=True
+        )
         mocker.patch.object(BaseValidator, "check_file_flags", return_value="")
         mocker.patch.object(
             PackUniqueFilesValidator, "are_valid_files", return_value=""
